EduPlanner_skill-tree/requirements.txt
Flask>=2.2
python-dotenv>=1.0
google-generativeai>=0.2.0
Flask-Caching>=1.10
Flask-Limiter>=2.0
bleach>=6.0
requests>=2.0